+++ AGENTS.md
@@ -29,6 +29,7 @@
 - **ユーザー入力は必ずバリデーション**: フォーム/URL/リクエストボディ等は必ず検証し、未検証の値をDBや外部APIに渡さない。
 - **変更を加えたら必ず型チェック**: 変更後は `npm run type-check` を実行して型の整合性を確認する。
 - **分割と構造化**: 可能な限りコンポーネント化し、200行以上のコードは分割を検討する。ディレクトリは責務ごとに分け、`schema` 層や `utils` 層などのまとまりで実装する。
+- **命名規則**: `parsed` のような曖昧な名前は避け、用途が分かる名称を使う（例: `*Validation`, `*Data`, `*Value`）。
 
 ## コマンド
 ```bash
