+++ app/interviews/_components/interview-fields.tsx
@@ -1,0 +1,62 @@
+type FieldProps = {
+  label: string;
+  value: string;
+  onChange: (value: string) => void;
+  placeholder?: string;
+  required?: boolean;
+};
+
+export type CompanyOption = { value: string; label: string };
+
+export function Field({ label, value, onChange, placeholder, required }: FieldProps) {
+  return (
+    <label className="block space-y-2">
+      <span className="text-sm font-medium text-slate-700 dark:text-slate-200">
+        {label}
+        {required ? <span className="text-rose-500"> *</span> : null}
+      </span>
+      <input
+        value={value}
+        onChange={(e) => onChange(e.target.value)}
+        required={required}
+        placeholder={placeholder}
+        className="w-full rounded-xl border border-slate-200 bg-white px-4 py-3 text-sm text-slate-900 outline-none focus:border-amber-300 dark:border-slate-700 dark:bg-slate-800 dark:text-slate-100"
+      />
+    </label>
+  );
+}
+
+type SelectFieldProps = {
+  label: string;
+  value: string;
+  onChange: (value: string) => void;
+  options: CompanyOption[];
+  required?: boolean;
+  placeholder?: string;
+};
+
+export function SelectField({ label, value, onChange, options, required, placeholder }: SelectFieldProps) {
+  return (
+    <label className="block space-y-2">
+      <span className="text-sm font-medium text-slate-700 dark:text-slate-200">
+        {label}
+        {required ? <span className="text-rose-500"> *</span> : null}
+      </span>
+      <select
+        value={value}
+        onChange={(e) => onChange(e.target.value)}
+        required={required}
+        className="w-full rounded-xl border border-slate-200 bg-white px-4 py-3 text-sm text-slate-900 outline-none focus:border-amber-300 dark:border-slate-700 dark:bg-slate-800 dark:text-slate-100"
+      >
+        <option value="" disabled className="text-slate-400">
+          {placeholder || "選択してください"}
+        </option>
+        {options.map((option) => (
+          <option key={option.value} value={option.value}>
+            {option.label}
+          </option>
+        ))}
+      </select>
+    </label>
+  );
+}
